+++ lib/api/subscribe.ex
@@ -1,17 +1,28 @@
 defmodule ExTier.Api.Subscribe do
-  alias ExTier.{Client, Phase}
+  alias ExTier.{Client, OrgInfo, Phase}
 
-  @type subscribe_params :: %{
+  @type subscribe_features_params :: %{
           :org => String.t(),
           :features => Phase.features() | [Phase.features()],
-          :effective => DateTime.t()
+          :effective => DateTime.t(),
+          optional(:info) => OrgInfo.t()
         }
+
+  @type subscribe_features_info :: %{
+          :org => String.t(),
+          optional(:features) => Phase.features() | [Phase.features()],
+          optional(:effective) => DateTime.t(),
+          :info => OrgInfo.t()
+        }
+
+  @type subscribe_params :: subscribe_features_params() | subscribe_features_info()
 
   @doc """
   Subscribe an organization to a plan or a set of features
 
       :ok = ExTier.schedule(%{org: "org:org_id", features: ["plan:my_plan@0"]})
       :ok = ExTier.schedule(%{org: "org:org_id", features: ["feature:feature@plan:my_plan@0"]})
+      :ok = ExTier.schedule(%{org: "org:org_id", info: %{email: "org@example.com"}})
 
   """
   @spec subscribe(subscribe_params) :: :ok | {:error, String.t()}
@@ -28,6 +39,10 @@ defmodule ExTier.Api.Subscribe do
       %{org: params.org}
       |> Map.put(:phases, [phases])
 
+    Client.post("/subscribe", params)
+  end
+
+  def subscribe(params) do
     Client.post("/subscribe", params)
   end
 end