+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Next
 
 - Add `ExTier.lookup/1` endpoint
+- Add `info` property in `ExTier.subscribe/1` endpoint
+
 ## v0.0.2
 
 - Add `ExTier.whoami/0` endpoint